working scanner change
