+++ flan/package.json
@@ -15,6 +15,7 @@
     "@skeletonlabs/skeleton": "2.0.0",
     "@skeletonlabs/tw-plugin": "0.1.0",
     "@sveltejs/adapter-auto": "^2.0.0",
+    "@sveltejs/adapter-static": "^2.0.3",
     "@sveltejs/kit": "^1.20.4",
     "@tailwindcss/typography": "0.5.10",
     "@types/node": "20.5.9",
