+++ flan/src/lib/QRCode.svelte
