+++ flan/src/routes/+layout.ts
@@ -1,0 +1,3 @@
+// This disables ssr on ALL pages (since its a layout)
+//https://kit.svelte.dev/docs/single-page-apps
+export const ssr = false;
