+++ flan/src/routes/+page.svelte
@@ -1,31 +1,10 @@
 <script lang="ts">
-	import { Html5QrcodeScanner, type QrcodeErrorCallback, type QrcodeSuccessCallback } from 'html5-qrcode';
 	import { goto } from '$app/navigation';
 	import { onMount } from 'svelte';
 	import { AppBar, FileDropzone } from '@skeletonlabs/skeleton';
-	import QRCode from '../components/QRCode.svelte';
-	import Logo from '../lib/Logo.svelte';
+	import QRCode from '$lib/QRCode.svelte';
+	import Logo from '$lib/Logo.svelte';
 	let link = '';
-	let scanned = '';
-	onMount(() => {
-		const onScanSuccess: QrcodeSuccessCallback = (decodedText, decodedResult) => {
-			if (decodedText == "") return
-			console.log(`Code matched = ${decodedText}`, decodedResult);
-			scanned = decodedText;
-			console.log("GO CHAT");
-		};
-
-		const onScanFailure: QrcodeErrorCallback = (error) => {
-			// console.warn(`Code scan error = ${error}`);
-		}
-
-		let html5QrcodeScanner = new Html5QrcodeScanner(
-			'scanner',
-			{ fps: 100, qrbox: { width: 250, height: 250 } },
-			/* verbose= */ false
-		);
-		html5QrcodeScanner.render(onScanSuccess, onScanFailure);
-	});
 </script>
 
 <AppBar gridColumns="grid-cols-3" slotDefault="place-self-center" slotTrail="place-content-end">
@@ -44,6 +23,5 @@
 		<input class="input" type="text" placeholder="Input" bind:value={link} />
 		<h1 class="h1">JOIN</h1>
 		<p class="h1">you join someone</p>
-		<div id="scanner" class="w-[600px]" />
 	</div>
 </div>
